# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Raffle Draw",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/customer/check-balance": {
            "post": {
                "tags": [
                    "Mobile"
                ],
                "operationId": "fd526b74450d24e1c51d0fa05309c908",
                "parameters": [
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "amount",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Balance is greater than amount",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/get-wins": {
            "get": {
                "tags": [
                    "General"
                ],
                "operationId": "6b21ff7aa529c7a1db656e896d8b1616",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "user_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Getting Stake successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/get-stakes": {
            "get": {
                "tags": [
                    "General"
                ],
                "operationId": "0b2ff4608d269ccb7e7eef95febc6994",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "user_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Getting Stake successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/customer/get-history": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List all the transactions",
                "operationId": "d46b261713bd20e5ed30d6e531e43cc9",
                "responses": {
                    "200": {
                        "description": "List all the transactions"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "List all winners by Date",
                "operationId": "ca29e6671706fa0bf6bfb5884def6da2",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start Date",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End Date",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List all winners by Date"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/v1/customer/get-debits-history": {
            "get": {
                "tags": [
                    "Mobile"
                ],
                "summary": "List debit transactions",
                "operationId": "5a7805a8ca2c223d469fd2d4404d4103",
                "responses": {
                    "200": {
                        "description": "List debit transactions"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/customer/get-credit-history": {
            "get": {
                "tags": [
                    "Mobile"
                ],
                "operationId": "219b23155faaea01e2111a3b75c35517",
                "responses": {
                    "200": {
                        "description": "Registration successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/customer/get-history": {
            "get": {
                "tags": [
                    "Mobile"
                ],
                "summary": "List Credit transactions",
                "operationId": "1a75a42fb026d50361ce9aab9ad63f4f",
                "responses": {
                    "200": {
                        "description": "List transactions"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/customer/get-withdrawal": {
            "get": {
                "tags": [
                    "Mobile"
                ],
                "summary": "Get Withdrawal",
                "operationId": "e59587b5bb140524379081ba9ba4fb12",
                "responses": {
                    "200": {
                        "description": "Getting withdrawal Sucessful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "User Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}